dom/base/test/test_tor_bug15502.html

<!DOCTYPE HTML>
<html>
<!--
https://bugs.torproject.org/15502
-->
<head>
  <meta http-equiv="content-type" content="text/html; charset=utf-8">
  <title>Test for Tor Browser Bug 15502</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript;version=1.7" src="bug15502_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content"></div>

<script class="testbody" type="application/javascript;version=1.7">
SimpleTest.waitForExplicitFinish();

// __prefs__. Import the `Serivces.prefs` object.
let prefs = SpecialPowers.Cu.import("resource://gre/modules/Services.jsm").Services.prefs;

// ## Testing constants
let domain1 = "http://example.com",
    domain2 = "http://example.net",
    path = "/tests/dom/base/test/",
    page_blob = "bug15502_page_blobify.html",
    page_deblob = "bug15502_page_deblobify.html"
    worker_blob = "bug15502_worker_blobify.html",
    worker_deblob = "bug15502_worker_deblobify.html";

// __tabIO(domain, child, input)__.
// Open a tab at the given `domain` and `child` page. Post an
// `input` message to the tab.
let tabIO = function* (domain, child, input) {
  tab = window.open(domain + path + "bug15502_tab.html", "_blank");
  yield receiveMessage(tab); // ready message
  sendMessage(tab, "http://example.org" + path + child);
  yield receiveMessage(tab); // ready message
  sendMessage(tab, input);
  return yield receiveMessage(tab);
};

// __blobTest(isolationOn, domainA, domainB, blobPage, deblobPage)__.
// Run a test where we set the pref "privacy.thirdparty.isolate` to on or off,
// and then create a blob URL in `domainA`, using the page `blobPage`,
// and then attempt to retrieve the object from the blobURL in `domainB`, using
// the page `deblobPage`.
let blobTest = function* (isolationOn, domainA, domainB, blobPage, deblobPage) {
  prefs.setIntPref("privacy.thirdparty.isolate", isolationOn ? 2 : 0);
  let input = "" + Math.random(),
      blobURL = yield tabIO(domainA, blobPage, input),
      result = yield tabIO(domainB, deblobPage, blobURL),
      description = domainA + ":" + blobPage + "->" + domainB + ":" + deblobPage + ", isolation " + (isolationOn ? "on." : "off.");
  if (blobPage === worker_blob) {
    // Remove this case when we write a patch that properly isolates web worker blob URLs
    // by first party domain.
    ok(blobURL.contains("Permission to call 'URL.createObjectURL' denied."), description + " Deny blob URL creation in web worker");
  } else if (deblobPage === worker_deblob && isolationOn) {
    // Remove this case when we write a patch that properly isolates web worker blob URLs
    // by first party domain.
    ok(result.contains("Access to restricted URI denied"), description + " Isolated blobs not available to web workers");
  } else {
    if (isolationOn && domainA !== domainB) {
      ok(input !== result, description + " Deny retrieval");
    } else {
      ok(input === result, description + " Allow retrieval");
    }
  }
};


// ## The main test
// Run a Task.jsm coroutine that tests various combinations of domains
// methods, and isolation states for reading and writing blob URLs.
Task.spawn(function* () {
  for (let isolate of [false, true]) {
    for (let domainB of [domain1, domain2]) {
      for (let blob of [page_blob, worker_blob]) {
        for (let deblob of [page_deblob, worker_deblob]) {
          yield blobTest(isolate, domain1, domainB, blob, deblob);
        }
      }
    }
  }
  SimpleTest.finish();
});

</script>

</body>
</html>
